+++ lib/landlock/native.rb
@@ -31,8 +31,20 @@ module Landlock
       Landlock.__send__(:_close_fd, fd)
     end
 
+    def close_inherited_fds!
+      Landlock.__send__(:_close_inherited_fds)
+    end
+
     def pidfd_open(pid)
       Landlock.__send__(:_pidfd_open, pid)
+    end
+
+    def arm_parent_death_process_group!(parent_pid)
+      Landlock.__send__(:_arm_parent_death_process_group, parent_pid)
+    end
+
+    def set_parent_death_signal!
+      Landlock.__send__(:_set_parent_death_signal)
     end
 
     def seccomp_deny_network!