+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- Add `Landlock.fork` for running a sandboxed Ruby block in a supervised forked child.
+
 ## [0.4.1] - 2026-08-20
 
 ### Fixed